requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
Flask-Mail==0.9.1
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
Flask-Login==0.6.3
razorpay
